requirements.txt
streamlit>=1.20
faiss-cpu>=1.7.4
numpy>=1.23
PyPDF2>=3.0
python-dotenv>=1.0.0
langchain-core>=0.0.5    # for the message classes used in app.py
requests>=2.28
tqdm>=4.65
pdfplumber
groq
langchain-groq
pypdf
sentence-transformers
langchain-text-splitters